requirements.txt
altair==5.5.0
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
ffmpeg-python==0.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.46.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.3.0
protobuf==6.31.1
pyarrow==20.0.0
PyAudio==0.2.14
pydeck==0.9.1
pydub==0.25.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
